requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.17
email-validator==2.2.0
